rsy/code/stockxWeb/requirements.txt
Flask==2.2.5
loguru==0.5.3
pandas==2.1.4
python_dateutil==2.8.2
pytz==2023.3.post1
openpyxl==3.0.10
pymysql==1.0.2
datetime
tqdm